backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.5.1
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.52.2
PyJWT==2.9.0

